Correct expected hashes in WyHashTest
I was copying the literal values out of the vector<uint64_t>,
just assuming those were the expected hashes.
However, that is not the case!

Instead, you have to run the actual `test_vector` program
to get the hash outputs.

I love it when it's the tests that have an error ;)